se agregan pruebas unitarias a la capa de servicio de User, tambien correccion a la clase Article y se añade Authcontroller

+++ src/main/java/com/primer_parcial8/Models/Article.java
@@ -31,10 +31,10 @@ public class Article {
     private int stock;
 
     @Column(nullable = false)
-    private Float priceSale;
+    private double priceSale;
 
     @Column(nullable = false)
-    private Float priceBuy;
+    private double priceBuy;
 
     @ManyToOne
     private Category  category;
